move get workers into workers controller
